Obj parser & storage fully implemented

+++ kit361-c/client.h
@@ -55,10 +55,10 @@ struct iMatrix
 // Point Storage
 struct point
 {
-	int x;
-	int y;
-	int z;
-	int w;
+	double x;
+	double y;
+	double z;
+	double w;
 };
 
 // Storage of Vertex
